+++ annofabcli/statistics/table.py
@@ -271,92 +271,3 @@ class Table:
             dtypes = {e: "float64" for e in numeric_columns}
             dtypes.update({e: "boolean" for e in boolean_columns})
             return pandas.DataFrame([], columns=columns).astype(dtypes)
-
-    @staticmethod
-    def create_annotation_count_ratio_df(
-        task_history_df: pandas.DataFrame, task_df: pandas.DataFrame
-    ) -> pandas.DataFrame:
-        """
-        task_id, phase, (phase_index), user_idの作業時間比から、アノテーション数などの生産量を求める
-
-        Args:
-
-        Returns:
-
-
-        """
-        annotation_count_dict = {
-            row["task_id"]: {
-                "annotation_count": row["annotation_count"],
-                "input_data_count": row["input_data_count"],
-                "inspection_comment_count": row["inspection_comment_count"],
-                "rejected_count": row["number_of_rejections_by_inspection"] + row["number_of_rejections_by_acceptance"],
-            }
-            for _, row in task_df.iterrows()
-        }
-
-        def get_inspection_comment_count(row: pandas.Series) -> float:
-            task_id = row.name[0]
-            inspection_comment_count = annotation_count_dict[task_id]["inspection_comment_count"]
-            return row["task_count"] * inspection_comment_count
-
-        def get_rejected_count(row: pandas.Series) -> float:
-            task_id = row.name[0]
-            rejected_count = annotation_count_dict[task_id]["rejected_count"]
-            return row["task_count"] * rejected_count
-
-        def get_annotation_count(row: pandas.Series) -> float:
-            task_id = row.name[0]
-            annotation_count = annotation_count_dict[task_id]["annotation_count"]
-            result = row["task_count"] * annotation_count
-            return result
-
-        def get_input_data_count(row: pandas.Series) -> float:
-            task_id = row.name[0]
-            annotation_count = annotation_count_dict[task_id]["input_data_count"]
-            return row["task_count"] * annotation_count
-
-        if len(task_df) == 0:
-            logger.warning("タスク一覧が0件です。")
-            return pandas.DataFrame()
-
-        group_obj = task_history_df.groupby(["task_id", "phase", "phase_stage", "account_id"]).agg(
-            {"worktime_hour": "sum"}
-        )
-        # 担当者だけ変更して作業していないケースを除外する
-        group_obj = group_obj[group_obj["worktime_hour"] > 0]
-
-        if len(group_obj) == 0:
-            logger.warning("タスク履歴情報に作業しているタスクがありませんでした。タスク履歴全件ファイルが更新されていない可能性があります。")
-            return pandas.DataFrame(
-                columns=[
-                    "task_id",
-                    "phase",
-                    "phase_stage",
-                    "account_id",
-                    "worktime_hour",
-                    "task_count",
-                    "annotation_count",
-                    "input_data_count",
-                    "pointed_out_inspection_comment_count",
-                    "rejected_count",
-                ]
-            )
-
-        group_obj["task_count"] = group_obj.groupby(level=["task_id", "phase", "phase_stage"], group_keys=False).apply(
-            lambda e: e / e["worktime_hour"].sum()
-        )
-        group_obj["annotation_count"] = group_obj.apply(get_annotation_count, axis="columns")
-        group_obj["input_data_count"] = group_obj.apply(get_input_data_count, axis="columns")
-        group_obj["pointed_out_inspection_comment_count"] = group_obj.apply(
-            get_inspection_comment_count, axis="columns"
-        )
-        group_obj["rejected_count"] = group_obj.apply(get_rejected_count, axis="columns")
-        new_df = group_obj.reset_index()
-        new_df["pointed_out_inspection_comment_count"] = new_df["pointed_out_inspection_comment_count"] * new_df[
-            "phase"
-        ].apply(lambda e: 1 if e == TaskPhase.ANNOTATION.value else 0)
-        new_df["rejected_count"] = new_df["rejected_count"] * new_df["phase"].apply(
-            lambda e: 1 if e == TaskPhase.ANNOTATION.value else 0
-        )
-        return new_df
